+++ src/index.ts
@@ -1,4 +1,5 @@
+export type { ByteChunk } from './fetch'
 export { parseText } from './text'
 export type { Newline, ParseError, ParseMeta, ParseOptions, ParseResult, State } from './types'
 export { parseURL, type ParseURLOptions } from './url'
-export { isEmptyLine, toURL } from './utils'
+export { type BlobURL, isEmptyLine, toBlobURL } from './utils'